requirements_andrei.txt
array_api_compat==1.8
certifi==2024.7.4
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
Cython==0.29.30
dataclasses==0.6
fonttools==4.53.1
fuzzywuzzy==0.18.0
idna==3.7
importlib_resources==6.4.0
joblib==1.4.2
kiwisolver==1.4.5
llvmlite==0.43.0
lxml==5.2.2
matplotlib==3.4.0
multidict==6.0.5
natsort==8.4.0
networkx==3.2.1
numpy==1.20.0
packaging==24.1
pandas==1.3.5
parso==0.8.4
pexpect==4.9.0
Pillow==8.4.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
protobuf==5.27.3
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow-hotfix==0.6
pydantic_core==2.20.1
Pygments==2.18.0
pyparsing==3.1.2
PyTDC==0.3.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
rdkit-pypi==2020.9.5.2
requests==2.32.3
scikit-learn==0.22
scipy==1.8.0
seaborn==0.10.0
six==1.16.0
soupsieve==2.5
stdlib-list==0.10.0
Theano==0.8.0
threadpoolctl==3.5.0
tomli==2.0.1
torch==1.8.0
tqdm==4.66.5
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.19
wcwidth==0.2.13
widgetsnbextension==4.0.11
wrapt==1.16.0
zipp==3.19.2
